Add a basic GUI.
- Spawn a very simple "explorer" that list the models loaded by the

application;
- Slight changes of the mem_allocator interface;
- Add the string data structure to the stdlib;
- Fix some bugs into the sse math library.

+++ src/renderer/regular/rdr_mesh.c
@@ -187,7 +187,7 @@ register_mesh_attribs
       goto error;
     }
 
-    assert(attr_usage < RDR_NB_ATTRIB_USAGES);
+    assert((int)attr_usage < RDR_NB_ATTRIB_USAGES);
     mesh->attrib_list[attr_usage].stride = mesh->vertex_size;
     mesh->attrib_list[attr_usage].offset = offset;
     mesh->attrib_list[attr_usage].type = rbattr_type;
